Move dev into JS folder
