+++ packages/javascript/dev/browser/backups/index.ts
@@ -1,5 +1,5 @@
 // import './antlr';
-import Coder from '../../../packages/javascript/autogrammer/src/index.js';
+import Coder from '../../../autogrammer/src/index.js';
 
 const coder = new Coder({
   model: 'Xenova/gpt2',
