+++ packages/javascript/dev/browser/global.d.ts
