+++ packages/javascript/dev/browser/package.json
