+++ packages/javascript/dev/browser/synthesize/index.html
