+++ packages/javascript/dev/browser/synthesize/index.ts
