+++ packages/javascript/dev/browser/vite.config.ts
