+++ packages/javascript/dev/node/src/commands/global.d.ts
